fix(sql) `wasm__invoke_function_*` returns `int8`.
Those functions were returning `int4` instead of `int8` (which is the
generic bit representation we use to represent all values for the moment).

Also, this patch replaces `integer` and `bigint`, resp. by `int4` and
`int8`. It's the same but more clear.

+++ src/foreign_data_wrappers/exported_functions.rs
@@ -35,8 +35,8 @@ impl ForeignData for ExportedFunctionsForeignDataWrapper {
         #[inline]
         fn wasm_type_to_pg_type(ty: &Type) -> &str {
             match ty {
-                Type::I32 => "integer",
-                Type::I64 => "bigint",
+                Type::I32 => "int4",
+                Type::I64 => "int8",
                 Type::F32 | Type::F64 => "numeric",
                 Type::V128 => "decimal",
             }
